fix(solid)!: import `SOLID_CORE_MAX` through `SOLID_CORE_MAX_VALUE` defined in `abi.hpp`
Fixes autocxx failure when `SOLID_CORE_MAX` is given by `BUILD_CFLAGS`.
It appears autocxx can't `generate!` a preprocessor definition injected
by clang args.

`abi::SOLID_CORE_MAX` is now of type `usize`.

+++ common/solid/src/smp.rs
@@ -15,7 +15,7 @@ pub fn current_processor_id() -> usize {
 /// Get the number of processors in the system.
 #[inline]
 pub fn num_processors() -> usize {
-    abi::SOLID_CORE_MAX as usize
+    abi::SOLID_CORE_MAX
 }
 
 #[derive(Clone, PartialEq, Eq, Debug, Hash)]
